Fix dragging header width
Attempting to change the width of a column causes the header view to
grow and shrink erratically. After investigating it appears the
event.pageX value used the first time SC.TableHeaderView:mouseDragged()
is called is undefined causing the newWidth calculation to generate
erratic values.

Saved the X position in the mouseDown() function then check the
event.pageX for undefined in mouseDragged(). If undefined use the saved
X position.